Create ATLANTIC MAMMALS: MEDIUM AND LARGE-SIZED
Biodiversity inventories contain important information about species richness, community structure and composition, and are the first step in developing any conservation and mitigation strategies. The Atlantic Forest of South America is home to around 334 species of small, medium and large-sized mammals, and is currently restricted to less than 12% of its original cover. Here we present the ATLANTIC MAMMALS, an open dataset on information on medium and large-sized mammal assemblages in the Atlantic Forest of Brazil, Paraguay and Argentina. A total of 129 studies were compiled, including published and in press peer-reviewed papers, book chapters, theses and unpublished data. We mapped 244 assemblages, eight orders, 63 genera and 94 species (24 of which are classified as threatened by the IUCN Red List) distributed in 128 protected and 116 unprotected areas. Species richness of the mammalian assemblages varied from 1 to 39 species (mean 15). The most recorded species in the entire biome was Dasypus novemcinctus, followed by Cerdocyon thous and Procyon cancrivorous. These data can be useful in support of macroecological studies and conservation planning strategies.